requirements.txt
streamlit>=1.20
google-genai>=0.14.0
pillow
python-dotenv
requests
pandas
numpy
